refactor : 토큰관련 코드 계층화(4)
