+++ src/main/java/com/kakaotech/team14backend/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.kakaotech.team14backend.config;
 
-import com.kakaotech.team14backend.filter.FilterResponseUtils;
+import com.kakaotech.team14backend._core.FilterResponseUtils;
 import com.kakaotech.team14backend._core.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
