+++ src/main/java/com/kakaotech/team14backend/oauth2/application/command/connect/GetInstagramAccessToken.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.oauth2.application.command;
+package com.kakaotech.team14backend.oauth2.application.command.connect;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
