+++ src/main/java/com/kakaotech/team14backend/oauth2/application/command/login/AuthenticationSuccessHandler.java
@@ -1,9 +1,10 @@
-package com.kakaotech.team14backend.oauth2.application.command;
+package com.kakaotech.team14backend.oauth2.application.command.login;
 
+import com.kakaotech.team14backend.oauth2.application.command.token.SetRefreshTokenCookie;
 import com.kakaotech.team14backend.oauth2.dto.TokenDTO;
 import com.kakaotech.team14backend.member.domain.Member;
 import com.kakaotech.team14backend.member.infrastructure.MemberRepository;
-import com.kakaotech.team14backend.oauth2.application.usecase.CreateOrUpdateTokens;
+import com.kakaotech.team14backend.oauth2.application.usecase.token.CreateOrUpdateTokens;
 import com.kakaotech.team14backend.oauth2.dto.KakaoProfileDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
