+++ src/main/java/com/kakaotech/team14backend/oauth2/application/command/login/CreateKakaoUser.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.oauth2.application.command;
+package com.kakaotech.team14backend.oauth2.application.command.login;
 
 
 import com.kakaotech.team14backend.oauth2.domain.PrincipalDetails;
