+++ src/main/java/com/kakaotech/team14backend/oauth2/application/command/token/CreateAccessToken.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.oauth2.application.command;
+package com.kakaotech.team14backend.oauth2.application.command.token;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
@@ -16,8 +16,12 @@ public class CreateAccessToken {
   private Long accessEXP;
   @Value("${jwt.token-validity-in-seconds-refreshToken}")
   private Long refreshEXP;
-  @Value("${jwt.secret}")
   private static String SECRET;
+
+  @Value("${jwt.secret}")
+  public void setSecret(String secret) {
+    SECRET = secret;
+  }
   public static String TOKEN_PREFIX = "Bearer ";
