+++ src/main/java/com/kakaotech/team14backend/oauth2/presentation/JwtController.java
@@ -4,9 +4,8 @@ package com.kakaotech.team14backend.oauth2.presentation;
 import com.kakaotech.team14backend.common.ApiResponse;
 import com.kakaotech.team14backend.common.ApiResponseGenerator;
 import com.kakaotech.team14backend.common.CookieUtils;
-import com.kakaotech.team14backend.oauth2.application.usecase.ReissueAccessTokenService;
+import com.kakaotech.team14backend.oauth2.application.usecase.token.ReissueAccessTokenService;
 import com.kakaotech.team14backend.oauth2.dto.ReissueDTO;
-import com.kakaotech.team14backend.jwt.service.TokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
