+++ src/main/java/com/kakaotech/team14backend/oauth2/presentation/LoginController.java
@@ -1,12 +1,12 @@
 package com.kakaotech.team14backend.oauth2.presentation;
 
-import com.kakaotech.team14backend.auth.PrincipalDetails;
+import com.kakaotech.team14backend.oauth2.domain.PrincipalDetails;
 import com.kakaotech.team14backend.common.ApiResponse;
 import com.kakaotech.team14backend.common.ApiResponseGenerator;
 import com.kakaotech.team14backend.common.MessageCode;
 import com.kakaotech.team14backend.member.exception.UserNotAuthenticatedException;
-import com.kakaotech.team14backend.oauth2.application.command.AuthenticationSuccessHandler;
-import com.kakaotech.team14backend.oauth2.application.usecase.LoginService;
+import com.kakaotech.team14backend.oauth2.application.command.login.AuthenticationSuccessHandler;
+import com.kakaotech.team14backend.oauth2.application.usecase.login.LoginService;
 import com.kakaotech.team14backend.oauth2.application.usecase.LogoutService;
 import com.kakaotech.team14backend.oauth2.dto.GetKakaoCodeDTO;
 import lombok.RequiredArgsConstructor;
