+++ src/main/java/com/kakaotech/team14backend/post/presentation/PostController.java
@@ -1,6 +1,6 @@
 package com.kakaotech.team14backend.post.presentation;
 
-import com.kakaotech.team14backend.auth.PrincipalDetails;
+import com.kakaotech.team14backend.oauth2.domain.PrincipalDetails;
 import com.kakaotech.team14backend.common.ApiResponse;
 import com.kakaotech.team14backend.common.ApiResponse.CustomBody;
 import com.kakaotech.team14backend.common.ApiResponseGenerator;
